+++ Bookworm/app/src/main/java/com/example/bookworm/SignUpActivity.java
@@ -3,7 +3,6 @@ package com.example.bookworm;
 import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
@@ -13,15 +12,10 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class SignUpActivity extends AppCompatActivity {
     public static final String TAG = "TAG";
@@ -98,19 +92,7 @@ public class SignUpActivity extends AppCompatActivity {
                                             .show();
 
                                     // Store user information
-                                    uid = fAuth.getCurrentUser().getUid();
-                                    DocumentReference documentReference = fStore.collection("users").document(uid);
-                                    Map<String, Object> userInfo = new HashMap<String, Object>();
-                                    userInfo.put("username", username);
-                                    userInfo.put("phoneNumber", phoneNumber);
-                                    userInfo.put("email", email);
-                                    documentReference.set(userInfo)
-                                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                                @Override
-                                                public void onSuccess(Void aVoid) {
-                                                    Log.d(TAG, "User profile is created for " + username);
-                                                }
-                                            });
+                                    Database.createUser(username, phoneNumber, email);
                                     startActivity(new Intent(getApplicationContext(), MainActivity.class));
                                 } else {
                                     Toast.makeText(SignUpActivity.this,
